# mcp-vocabulary/api/_generated/mcp_vocabulary_api/models/add_vocabulary_request.py
# coding: utf-8

"""
    Kapampangan Vocabulary MCP

    Vocabulary lookup service for the Kapampangan language tutor. Uses pgvector cosine similarity search against a PostgreSQL vocabulary table populated from canonical data files.  Called by the orchestration layer's agentic tool loop. Can also be queried directly here for manual testing. 

    The version of the OpenAPI document: 0.2.0
    Generated by OpenAPI Generator (https://openapi-generator.tech)

    Do not edit the class manually.
"""  # noqa: E501


from __future__ import annotations
import pprint
import re  # noqa: F401
import json




from pydantic import BaseModel, ConfigDict, Field, StrictStr
from typing import Any, ClassVar, Dict, List, Optional
from typing_extensions import Annotated
try:
    from typing import Self
except ImportError:
    from typing_extensions import Self

class AddVocabularyRequest(BaseModel):
    """
    AddVocabularyRequest
    """ # noqa: E501
    term: StrictStr = Field(description="Kapampangan word or phrase")
    meaning: StrictStr = Field(description="English meaning")
    part_of_speech: Optional[StrictStr] = Field(default=None, description="verb | noun | adjective | phrase")
    aspect_forms: Optional[Dict[str, StrictStr]] = Field(default=None, description="For verbs: {\"progressive\": \"...\", \"completed\": \"...\", \"contemplated\": \"...\"}")
    examples: Optional[List[Dict[str, Any]]] = Field(default=None, description="[{\"kapampangan\": \"...\", \"english\": \"...\"}]")
    usage_notes: Optional[StrictStr] = None
    source: Optional[StrictStr] = Field(default=None, description="native_speaker | reference | inferred")
    authority_level: Optional[Annotated[int, Field(le=4, strict=True, ge=1)]] = Field(default=3, description="1=native speaker verified, 2=academic, 3=community, 4=LLM inferred")
    __properties: ClassVar[List[str]] = ["term", "meaning", "part_of_speech", "aspect_forms", "examples", "usage_notes", "source", "authority_level"]

    model_config = {
        "populate_by_name": True,
        "validate_assignment": True,
        "protected_namespaces": (),
    }


    def to_str(self) -> str:
        """Returns the string representation of the model using alias"""
        return pprint.pformat(self.model_dump(by_alias=True))

    def to_json(self) -> str:
        """Returns the JSON representation of the model using alias"""
        # TODO: pydantic v2: use .model_dump_json(by_alias=True, exclude_unset=True) instead
        return json.dumps(self.to_dict())

    @classmethod
    def from_json(cls, json_str: str) -> Self:
        """Create an instance of AddVocabularyRequest from a JSON string"""
        return cls.from_dict(json.loads(json_str))

    def to_dict(self) -> Dict[str, Any]:
        """Return the dictionary representation of the model using alias.

        This has the following differences from calling pydantic's
        `self.model_dump(by_alias=True)`:

        * `None` is only added to the output dict for nullable fields that
          were set at model initialization. Other fields with value `None`
          are ignored.
        """
        _dict = self.model_dump(
            by_alias=True,
            exclude={
            },
            exclude_none=True,
        )
        return _dict

    @classmethod
    def from_dict(cls, obj: Dict) -> Self:
        """Create an instance of AddVocabularyRequest from a dict"""
        if obj is None:
            return None

        if not isinstance(obj, dict):
            return cls.model_validate(obj)

        _obj = cls.model_validate({
            "term": obj.get("term"),
            "meaning": obj.get("meaning"),
            "part_of_speech": obj.get("part_of_speech"),
            "aspect_forms": obj.get("aspect_forms"),
            "examples": obj.get("examples"),
            "usage_notes": obj.get("usage_notes"),
            "source": obj.get("source"),
            "authority_level": obj.get("authority_level") if obj.get("authority_level") is not None else 3
        })
        return _obj