# mcp-vocabulary/api/_generated/mcp_vocabulary_api/models/admin_export_response.py
# coding: utf-8

"""
    Kapampangan Vocabulary MCP

    Vocabulary lookup service for the Kapampangan language tutor. Uses pgvector cosine similarity search against a PostgreSQL vocabulary table populated from canonical data files.  Called by the orchestration layer's agentic tool loop. Can also be queried directly here for manual testing. 

    The version of the OpenAPI document: 0.2.0
    Generated by OpenAPI Generator (https://openapi-generator.tech)

    Do not edit the class manually.
"""  # noqa: E501


from __future__ import annotations
import pprint
import re  # noqa: F401
import json




from pydantic import BaseModel, ConfigDict
from typing import Any, ClassVar, Dict, List
from mcp_vocabulary_api.models.vocabulary_entry import VocabularyEntry
try:
    from typing import Self
except ImportError:
    from typing_extensions import Self

class AdminExportResponse(BaseModel):
    """
    AdminExportResponse
    """ # noqa: E501
    vocabulary: List[VocabularyEntry]
    __properties: ClassVar[List[str]] = ["vocabulary"]

    model_config = {
        "populate_by_name": True,
        "validate_assignment": True,
        "protected_namespaces": (),
    }


    def to_str(self) -> str:
        """Returns the string representation of the model using alias"""
        return pprint.pformat(self.model_dump(by_alias=True))

    def to_json(self) -> str:
        """Returns the JSON representation of the model using alias"""
        # TODO: pydantic v2: use .model_dump_json(by_alias=True, exclude_unset=True) instead
        return json.dumps(self.to_dict())

    @classmethod
    def from_json(cls, json_str: str) -> Self:
        """Create an instance of AdminExportResponse from a JSON string"""
        return cls.from_dict(json.loads(json_str))

    def to_dict(self) -> Dict[str, Any]:
        """Return the dictionary representation of the model using alias.

        This has the following differences from calling pydantic's
        `self.model_dump(by_alias=True)`:

        * `None` is only added to the output dict for nullable fields that
          were set at model initialization. Other fields with value `None`
          are ignored.
        """
        _dict = self.model_dump(
            by_alias=True,
            exclude={
            },
            exclude_none=True,
        )
        # override the default output from pydantic by calling `to_dict()` of each item in vocabulary (list)
        _items = []
        if self.vocabulary:
            for _item in self.vocabulary:
                if _item:
                    _items.append(_item.to_dict())
            _dict['vocabulary'] = _items
        return _dict

    @classmethod
    def from_dict(cls, obj: Dict) -> Self:
        """Create an instance of AdminExportResponse from a dict"""
        if obj is None:
            return None

        if not isinstance(obj, dict):
            return cls.model_validate(obj)

        _obj = cls.model_validate({
            "vocabulary": [VocabularyEntry.from_dict(_item) for _item in obj.get("vocabulary")] if obj.get("vocabulary") is not None else None
        })
        return _obj